change gravity; add tools
