+++ simulation/main.py
@@ -5,32 +5,11 @@ import sys
 import os
 import pybullet as p
 from stretch import *
+from utils.tools import *
 
 p.connect(p.GUI)
 p.configureDebugVisualizer(p.COV_ENABLE_GUI, 1)
-p.setGravity(0, 0, -1.81)
-
-def get_joint_index_by_name(robot, joint_name):
-    num_joints = p.getNumJoints(robot)
-    
-    for i in range(num_joints):
-        joint_info = p.getJointInfo(robot, i)
-        if joint_info[1].decode("utf-8") == joint_name:
-            print(f"joint name: {joint_name} found at index {i}")
-            return i
-    return None  # Return None if the joint name is not found
-
-def get_link_index_by_name(robot, link_name):
-    num_joints = p.getNumJoints(robot)  # This gives the number of joints, which is one less than the number of links
-    
-    for i in range(num_joints):
-        link_info = p.getJointInfo(robot, i)
-        link_name_in_urdf = link_info[12].decode("utf-8")  # Link name is stored at index 12
-        if link_name_in_urdf == link_name:
-            print(f"link name: {link_name_in_urdf} found at index {i}")
-            return i
-    return None  # Return None if the link name is not found
-
+p.setGravity(0, 0, -9.81)
 
 mobot = init_scene(p)
     
@@ -122,6 +101,6 @@ while (1):
     base_control(mobot, p, forward, turn)
     arm_control(mobot, p, up, stretch, roll, yaw)
     gripper_control(mobot, p, gripper_open)
-   
+    
     mobot.get_observation()
